pyproject.toml
[project]
name = "teclusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.6",
    "openai>=1.95.1",
    "openai-agents>=0.2.0",
    "opencv-python>=4.12.0.88",
    "orjson>=3.11.1",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scipy>=1.16.1",
    "yt-dlp>=2025.7.21",
]
